requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-heroku==0.3.1
django-storages==1.7.1
djangorestframework==3.9.1
docutils==0.15.2
gunicorn==19.9.0
jmespath==0.9.4
Pillow==6.2.0
psycopg2==2.8.3
python-dateutil==2.8.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
django-filter==2.2.0
requests==2.22.0
django-polymorphic==2.1.2
django-rest-swagger==2.2.0